refactor(racingCar): Controller 클래스 생성
- Application 클래스에 할당되어 있는 역할, 책임을 Controller에 부여
